Auto-fuzz: Reorganise codes and python files
