--- tools/auto-fuzz/benchmark_target.py
@@ -1,35 +1,0 @@
-TARGET_METHOD = {
-    'jvm': {
-        'benchmark1': [
-            'public static java.lang.Boolean parseData(java.lang.String,java.lang.Integer,java.lang.Integer) throws autofuzz.benchmark.object.AutoFuzzException'
-        ],
-        'benchmark2': [
-            'public static java.lang.Boolean parseData(java.lang.String,java.lang.Integer,java.lang.Integer) throws autofuzz.benchmark.object.AutoFuzzException'
-        ],
-        'benchmark3': [
-            'public static java.lang.Boolean parseData(java.lang.String,java.lang.Integer,java.lang.Integer,java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException'
-        ],
-        'benchmark4': [
-            'public static java.lang.Boolean parseData(java.lang.String,java.lang.Integer) throws autofuzz.benchmark.object.AutoFuzzException',
-            'public static java.lang.Boolean parseData(java.lang.String,java.lang.Integer,java.lang.Integer,java.lang.Integer) throws autofuzz.benchmark.object.AutoFuzzException'
-        ],
-        'benchmark5': [
-            'public void parseAlphabetic(java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException',
-            'public void parseInteger(java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException',
-            'public void parseFloat(java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException'
-        ],
-        'benchmark6': [
-            'public void parseData(java.lang.String[]) throws autofuzz.benchmark.object.AutoFuzzException'
-        ],
-        'benchmark7': [
-            'public java.lang.Boolean processClass(java.lang.Class,java.lang.String,java.lang.Integer) throws autofuzz.benchmark.object.AutoFuzzException'
-        ],
-        'benchmark8': [
-            'public void entry1(java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException',
-            'public void entry2(java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException',
-            'public void entry3(java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException',
-            'public void entry4(java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException',
-            'public void entry5(java.lang.String) throws autofuzz.benchmark.object.AutoFuzzException'
-        ]
-    }
-}
